requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
distlib==0.3.8
Django==5.1.3
django-allauth==0.63.6
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.13.3
idna==3.7
inflection==0.5.1
packaging==24.2
pillow==10.4.0
pipenv==2023.12.1
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==69.2.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.25.1